Hide ProjectFile internals
