+++ src/main.c
@@ -24,6 +24,7 @@
 #include "preferences.c"
 #include "preferences_dialog.c"
 #include "process.c"
+#include "project_file.c"
 #include "project.c"
 #include "real_process.c"
 #include "real_swank_process.c"
