+++ tests/Makefile
@@ -22,10 +22,10 @@ swank_session_test: swank_session_test.c swank_session.c swank_process.c real_sw
 lisp_parser_test: lisp_parser_test.c lisp_lexer.c lisp_parser.c node.c package.c string_text_provider.c text_provider.c
 	$(CC) $(CFLAGS) $^ -o $@ $(LDLIBS)
 
-project_test: project_test.c project.c analyser.c node.c package.c lisp_lexer.c lisp_parser.c string_text_provider.c text_provider.c
+project_test: project_test.c project.c project_file.c analyser.c node.c package.c lisp_lexer.c lisp_parser.c string_text_provider.c text_provider.c
 	$(CC) $(CFLAGS) $^ -o $@ $(LDLIBS)
 
-asdf_test: asdf_test.c asdf.c string_text_provider.c text_provider.c lisp_lexer.c lisp_parser.c node.c package.c project.c analyser.c
+asdf_test: asdf_test.c asdf.c string_text_provider.c text_provider.c lisp_lexer.c lisp_parser.c node.c package.c project.c project_file.c analyser.c
 	$(CC) $(CFLAGS) $^ -o $@ $(LDLIBS)
 
 analyser_test: analyser_test.c analyser.c node.c package.c lisp_lexer.c lisp_parser.c string_text_provider.c text_provider.c
